feat: create products table migration
This commit adds a new migration file `2024_07_15_065710_create_products_table.php` that creates the `products` table in the database. The table has columns for `name` (string) and `price` (double), along with the default `id` and `timestamps` columns. This migration will be used to create the necessary table for the `Product` model.