chore: Update product-script.blade.php to include update_product_form click event
This commit adds a click event handler for the "update_product_form" class in the product-script.blade.php file. When the element with this class is clicked, it retrieves the ID, name, and price data attributes and sets the corresponding input fields in the updateProductModal. This allows users to easily update the details of a product.