chore: Update click event for adding a product in product-script.blade.php
This commit updates the click event for adding a product in the product-script.blade.php file. The event is now delegated to the document level using the "on" method, which ensures that the event handler is applied to dynamically added elements as well. This change allows the "btnAddProduct" button to work correctly even after the page has been modified.